+++ src/Budgeter.Domain/ResourceDelta.cs
@@ -4,8 +4,8 @@ namespace Budgeter.Domain
 {
     public class ResourceDelta: IEntity
     {
-        public int Id { get; set; }
-        public int CategoryId { get; set; }
+        public long Id { get; set; }
+        public long CategoryId { get; set; }
 
         public decimal Amount { get; set; }
         public string Description { get; set; }
